Fix issue #1738

+++ synfig-studio/src/gui/widgets/widget_soundwave.cpp
@@ -197,7 +197,7 @@ bool Widget_SoundWave::on_draw(const Cairo::RefPtr<Cairo::Context>& cr)
 		int value = middle_y;
 		if (t >= sound_delay) {
 			synfig::Time dt = t - sound_delay;
-			int index = int(dt * stride) + channel_idx;
+			unsigned long index = int(dt * stride) + channel_idx;
 			if (index >= 0 && index < buffer.size())
 				std::copy(buffer.begin() + index, buffer.begin() + index + bytes_per_sample, &value);
 		}
